feat:support async metadata transfer.
