+++ spring-cloud-tencent-commons/src/main/java/com/tencent/cloud/common/metadata/config/MetadataLocalProperties.java
@@ -52,10 +52,6 @@ public class MetadataLocalProperties {
 	 * A transitive http header key list.
 	 */
 	private List<String> headers;
-	/**
-	 * Async mode config.
-	 */
-	private Async async = new Async();
 
 	public Map<String, String> getContent() {
 		if (CollectionUtils.isEmpty(content)) {
@@ -101,14 +97,6 @@ public class MetadataLocalProperties {
 		this.headers = headers;
 	}
 
-	public Async getAsync() {
-		return async;
-	}
-
-	public void setAsync(Async async) {
-		this.async = async;
-	}
-
 	@Override
 	public String toString() {
 		return "MetadataLocalProperties{" +
@@ -116,33 +104,6 @@ public class MetadataLocalProperties {
 				", transitive=" + transitive +
 				", disposable=" + disposable +
 				", headers=" + headers +
-				", async=" + async +
 				'}';
-	}
-
-	/**
-	 * Async mode config.
-	 */
-	public static class Async {
-
-		/**
-		 * Enable async or not.
-		 */
-		private Boolean enabled = false;
-
-		public Boolean getEnabled() {
-			return enabled;
-		}
-
-		public void setEnabled(Boolean enabled) {
-			this.enabled = enabled;
-		}
-
-		@Override
-		public String toString() {
-			return "Async{" +
-					"enabled=" + enabled +
-					'}';
-		}
 	}
 }
